+++ zcash_client_backend/src/data_api/ll.rs
@@ -653,7 +653,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for WalletSaplingOutput<AccountId> 
     }
     fn transfer_type(&self) -> TransferType {
         if self.is_change() {
-            TransferType::WalletInternal
+            TransferType::AccountInternal
         } else {
             TransferType::Incoming
         }
@@ -694,7 +694,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for DecryptedOutput<::sapling::Note
         self.transfer_type()
     }
     fn is_change(&self) -> bool {
-        self.transfer_type() == TransferType::WalletInternal
+        self.transfer_type() == TransferType::AccountInternal
     }
     fn nullifier(&self) -> Option<&::sapling::Nullifier> {
         None
@@ -703,7 +703,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for DecryptedOutput<::sapling::Note
         None
     }
     fn recipient_key_scope(&self) -> Option<Scope> {
-        if self.transfer_type() == TransferType::WalletInternal {
+        if self.transfer_type() == TransferType::AccountInternal {
             Some(Scope::Internal)
         } else {
             Some(Scope::External)
@@ -744,7 +744,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for WalletOrchardOutput<AccountId> 
     }
     fn transfer_type(&self) -> TransferType {
         if self.is_change() {
-            TransferType::WalletInternal
+            TransferType::AccountInternal
         } else {
             TransferType::Incoming
         }
@@ -786,7 +786,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for DecryptedOutput<::orchard::Note
         self.transfer_type()
     }
     fn is_change(&self) -> bool {
-        self.transfer_type() == TransferType::WalletInternal
+        self.transfer_type() == TransferType::AccountInternal
     }
     fn nullifier(&self) -> Option<&::orchard::note::Nullifier> {
         None
@@ -795,7 +795,7 @@ impl<AccountId: Copy> ReceivedShieldedOutput for DecryptedOutput<::orchard::Note
         None
     }
     fn recipient_key_scope(&self) -> Option<Scope> {
-        if self.transfer_type() == TransferType::WalletInternal {
+        if self.transfer_type() == TransferType::AccountInternal {
             Some(Scope::Internal)
         } else {
             Some(Scope::External)